# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cheku Left API",
        "description": "API documentation for Cheku Left POS system",
        "contact": {
            "email": "support@chekuleft.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "List sales",
                "description": "Get list of sales with optional filters",
                "operationId": "e9071d925184ec3b9522775e6810a95f",
                "parameters": [
                    {
                        "name": "butcher_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sales"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create a sale",
                "description": "Record a new sale",
                "operationId": "9741700fd7f92aca7b46ae0fec83a5e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_sale_id",
                                    "total_amount",
                                    "payment_method",
                                    "sale_date",
                                    "items"
                                ],
                                "properties": {
                                    "device_sale_id": {
                                        "type": "string"
                                    },
                                    "sale_number": {
                                        "type": "string"
                                    },
                                    "total_amount": {
                                        "type": "number"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}